feat(trivy): simplify install

+++ src/trivy/install.sh
@@ -14,28 +14,17 @@ if [ -z "${VERSION:-}" ]; then
   VERSION="$(get_latest_gh_release aquasecurity/trivy)"
 fi
 
-case "$(uname -m)" in
-x86_64)
-  ARCHITECTURE="64bit"
-  ;;
-aarch64 | armv8* | arm64)
-  ARCHITECTURE="ARM64"
-  ;;
-*)
-  echo "Architecture unsupported"
-  exit 1
-  ;;
-esac
-
-FILENAME="trivy_${VERSION}_Linux-${ARCHITECTURE}.tar.gz"
-URL="https://github.com/aquasecurity/trivy/releases/download/v${VERSION}/${FILENAME}"
-curl -L "$URL" -o "/tmp/${FILENAME}"
-tar \
-  --extract \
-  --verbose \
-  --auto-compress \
-  --file="/tmp/${FILENAME}" \
-  --directory=/usr/local/bin \
-  trivy
-chmod 0755 /usr/local/bin/trivy
-rm -rf "/tmp/${FILENAME}"
+if ! trivy version | grep -q -e "$VERSION$"; then
+  FILENAME="trivy_${VERSION}_Linux-$(get_arch64_b).tar.gz"
+  URL="https://github.com/aquasecurity/trivy/releases/download/v${VERSION}/${FILENAME}"
+  curl -L "$URL" -o "/tmp/${FILENAME}"
+  tar \
+    --extract \
+    --verbose \
+    --auto-compress \
+    --file="/tmp/${FILENAME}" \
+    --directory=/usr/local/bin \
+    trivy
+  chmod 0755 /usr/local/bin/trivy
+  rm -rf "/tmp/${FILENAME}"
+fi
